pyproject.toml
[project]
name = "openweather-mcp-server"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server that provides weather information using the OpenWeather API with LangChain integrations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastmcp>=2.12.0",
    "requests>=2.32.5",
]